net: update False Start blacklist for 32-bit indexes.
(Needed for the next list update.)

BUG=none
TEST=net_unittests

http://codereview.chromium.org/4599002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65376 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/ssl_false_start_blacklist.cc
@@ -12,8 +12,8 @@ bool SSLFalseStartBlacklist::IsMember(const char* host) {
   if (!last_two_labels)
     return false;
   const unsigned bucket = Hash(last_two_labels) & (kBuckets - 1);
-  const uint16 start = kHashTable[bucket];
-  const uint16 end = kHashTable[bucket + 1];
+  const uint32 start = kHashTable[bucket];
+  const uint32 end = kHashTable[bucket + 1];
   const size_t len = strlen(host);
 
   for (size_t i = start; i < end;) {
